requirements.txt
alembic==1.17.2
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
cachetools==6.2.6
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.124.4
filelock==3.25.2
fonttools==4.61.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.68.0
googleapis-common-protos==1.73.1
greenlet==3.3.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.8.0
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
narwhals==2.17.0
networkx==3.6.1
numpy==2.4.2
ollama==0.6.1
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==12.1.1
plotly==6.5.2
pluggy==1.6.0
postgrest==2.25.1
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
psycopg2-binary==2.9.11
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.2
pyparsing==3.3.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytrends==4.9.2
pytz==2025.2
PyYAML==6.0.3
realtime==2.25.1
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.45
sqlmodel==0.0.38
starlette==0.50.0
storage3==2.25.1
streamlit==1.54.0
StrEnum==0.4.15
supabase==2.25.1
supabase-auth==2.25.1
supabase-functions==2.25.1
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.11.0
tornado==6.5.4
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.38.0
watchdog==6.0.0
websockets==15.0.1
wordcloud==1.9.6
yarl==1.22.0
